INTELLIJ-122 Fixed support for IntelliJ Community Edition
